pyproject.toml
[project]
name = "juma-agent-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "db-dtypes>=1.5.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-cloud-bigquery>=3.39.0",
    "google-genai>=1.56.0",
    "gunicorn>=23.0.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "tabulate>=0.9.0",
]
